chore: rename 'auth-api.js' to 'user-api.js'
